+++ 2020/day02/index.test.ts
@@ -1,21 +1,15 @@
-import { solver1, solver2 } from "../day08";
+import { solver1, solver2 } from ".";
 
-const sampleData = `nop +0
-acc +1
-jmp +4
-acc +3
-jmp -3
-acc -99
-acc +1
-jmp -4
-acc +6`;
+const sampleData = `1-3 a: abcde
+1-3 b: cdefg
+2-9 c: ccccccccc`;
 
-describe("day08", () => {
+describe("day02", () => {
     describe("part1", () => {
         it("should solve for sample data", () => {
             const result = solver1(sampleData);
 
-            expect(result).toEqual(5);
+            expect(result).toEqual(2);
         });
     });
 
@@ -23,7 +17,7 @@ describe("day08", () => {
         it("should solve for sample data", () => {
             const result = solver2(sampleData);
 
-            expect(result).toEqual(8);
+            expect(result).toEqual(1);
         });
     });
 });
