+++ 2020/day04/index.test.ts
@@ -1,4 +1,4 @@
-import { solver1, solver2 } from "../day04";
+import { solver1, solver2 } from ".";
 
 const sampleData = `ecl:gry pid:860033327 eyr:2020 hcl:#fffffd
 byr:1937 iyr:2017 cid:147 hgt:183cm
