+++ 2020/day10/index.test.ts
@@ -1,4 +1,4 @@
-import { solver1, solver2 } from "../day10";
+import { solver1, solver2 } from ".";
 
 const sampleData = `16
 10
