+++ 2020/day21/index.test.ts
@@ -1,0 +1,24 @@
+import { solver1, solver2 } from ".";
+
+const sampleData = `mxmxvkd kfcds sqjhc nhms (contains dairy, fish)
+trh fvjkl sbzzf mxmxvkd (contains dairy)
+sqjhc fvjkl (contains soy)
+sqjhc mxmxvkd sbzzf (contains fish)`;
+
+describe("day21", () => {
+    describe("part1", () => {
+        it("should solve for sample data", () => {
+            const result = solver1(sampleData);
+
+            expect(result).toEqual(5);
+        });
+    });
+
+    describe("part2", () => {
+        it("should solve for sample data", () => {
+            const result = solver2(sampleData);
+
+            expect(result).toEqual("mxmxvkd,sqjhc,fvjkl");
+        });
+    });
+});
